Move inline auth to Client from Logical
Keeping this on Logical() limits the consumption; uses like Sys() and
KVv2() and other top-level clients cannot also use this without
reimplementing it. By moving it to Client, it sits alongside
SetToken(...) and thus should be more discoverable.

As this has not landed in a release, it shouldn't require a changelog
entry.
