+++ api/client.go
@@ -6,6 +6,8 @@ package api
 import (
 	"context"
 	"crypto/tls"
+	"encoding/base64"
+	"encoding/json"
 	"errors"
 	"fmt"
 	"math"
@@ -92,6 +94,9 @@ const (
 		"    BAO_ADDR=http://<address> vault <command>\n\n" +
 		"where <address> is replaced by the actual address to the server."
 )
+
+// InlineAuthOpts represents an option for inline authentication
+type InlineAuthOpts func() map[string][]string
 
 // Deprecated values
 const (
@@ -1571,6 +1576,77 @@ func (c *Client) WithResponseCallbacks(callbacks ...ResponseCallback) *Client {
 	c2.modifyLock = sync.RWMutex{}
 	c2.responseCallbacks = callbacks
 	return &c2
+}
+
+// InlineWithNamespace is used with WithInlineAuth(...) to set the namespace
+// of the inline authentication call.
+func InlineWithNamespace(ns string) InlineAuthOpts {
+	return func() map[string][]string {
+		return map[string][]string{
+			InlineAuthNamespaceHeaderName: {ns},
+		}
+	}
+}
+
+// InlineWithOperation is used with WithInlineAuth(...) to set the operation
+// of the inline authentication call.
+func InlineWithOperation(op string) InlineAuthOpts {
+	return func() map[string][]string {
+		return map[string][]string{
+			InlineAuthOperationHeaderName: {op},
+		}
+	}
+}
+
+// WithInlineAuth returns a client with no authentication information but
+// which sets headers which perform inline authentication. This
+// re-authenticates on every request and does not persist any token.
+// Operations which result in lease creation will not work.
+//
+// Refer to the OpenBao documentation for more information.
+func (c *Client) WithInlineAuth(path string, data map[string]interface{}, opts ...InlineAuthOpts) (*Client, error) {
+	client, err := c.Clone()
+	if err != nil {
+		return nil, fmt.Errorf("error cloning client: %w", err)
+	}
+
+	headers := client.Headers()
+	for h := range client.Headers() {
+		if strings.HasPrefix(h, InlineAuthParameterHeaderPrefix) {
+			delete(headers, h)
+		}
+	}
+
+	delete(headers, InlineAuthOperationHeaderName)
+	delete(headers, InlineAuthNamespaceHeaderName)
+	delete(headers, AuthHeaderName)
+
+	headers[InlineAuthPathHeaderName] = []string{path}
+
+	for _, opt := range opts {
+		oHeader := opt()
+		for name, value := range oHeader {
+			headers[name] = value
+		}
+	}
+
+	for key, value := range data {
+		jEncoded, err := json.Marshal(map[string]interface{}{
+			"key":   key,
+			"value": value,
+		})
+		if err != nil {
+			return nil, fmt.Errorf("failed to encode inline auth data key `%v`: %w", key, err)
+		}
+
+		b64Encoded := base64.RawURLEncoding.EncodeToString(jEncoded)
+		headers[fmt.Sprintf("%v%v", InlineAuthParameterHeaderPrefix, key)] = []string{b64Encoded}
+	}
+
+	client.ClearToken()
+	client.SetHeaders(headers)
+
+	return client, nil
 }
 
 // withConfiguredTimeout wraps the context with a timeout from the client configuration.
